Move sample dm related functions closer to sample dm
